feat(RSVP): Refactor data/rsvp to re-export from shared store
Update V1 RSVP data layer to import from rsvp-shared:
- index.js: Re-export types, actions, selectors, sagas from shared
- reducer.js: Compose shared reducer with V1 header-image reducer
- sagas.js: Fork shared sagas + V1-specific header-image and post-save sagas
- thunks.js: Use shared actions and hydration utils
- header-image.js: Import type from shared store
- Add header-image-sagas.js and post-save-sagas.js (V1-specific sagas)
- Update sagas test imports

Part of SOFT-3335
