+++ src/modules/data/blocks/rsvp/index.js
@@ -1,12 +1,9 @@
 /**
  * Internal dependencies
  */
-import * as types from './types';
-import * as actions from './actions';
-import * as selectors from './selectors';
-import * as thunks from './thunks';
+export { types, actions, selectors, sagas } from '../rsvp-shared';
 import reducer from './reducer';
-import sagas from './sagas';
+import * as thunks from './thunks';
 
 export default reducer;
-export { types, actions, sagas, selectors, thunks };
+export { thunks };
